Create README.md
The website design is simple, straightforward, and responsive. It is suitable for any device be it a laptop, tablet, or mobile phone. It contains a header that includes three pages the home page, services page, and work page. The top right corner also has logos and links to various social media handles of the company.